Added database support. We have a table for users, with discord id, address and moniker.
Each user has a list of entries for missing block counts, with timestamps.

Simplified the Kujira and KujiraPool classes.

Kujira class now checks the health of the node by asking a list of validators from the server. If successful, then the node marked as active.

KujiraPool has a list of Kujira nodes. It updates the node health status, and when we ask for the missing block counts, selects one of the active ones randomly. If we have 20 active kujira nodes, it will asked randomly to one of them.

We are now retrieving the list of addresses from the database. Database will be populated via discord bot.

Used SQLAlchemy for database operations. By changing the engine string at models.py, we can change the database.